Fix for single-feature regression. (#596)
* Fix for single-feature regression.

After the latest set of merges, I found that a single feature
was not rendering on the map. The check for "source.data.features"
was not properly testing for the case where there could be a single
feature stored in "data".  I've updated teh check and it appears to
be working fine now.

* Added unit test for geojson source types.

1. Tests for undefined, empty, Feature, and FeatureCollection.
   These tests are all passing.
2. The URI test is failing as the feature is unimplemented.

* GeoJson source now handles a uri.

1. Updated tests to cover more situations for
   data input to the reducer.
2. Updated the map component to fetch a remote geojson file
   as required by the mapbox gl spec.