Grouping GL Layers into OL Layers for performance optimizations (#613)
* Map will now group layers by source.

If multiple layers are coming from the same source,
they will now be logically grouped and styles will be used
in order to control them.

* Updated the lib deps for MVT styles and added example.

* Cleaned up the examples lint + prevent conf.js add

1. Fixed all the lint issues for the examples.
2. Created a new gitignore rule to allow placing a conf.js
   file in the examples for adding mapbox access tokens.
3. Removed old tests for sprites that are no handled
   by ol-mapbox-style.
4. added success test to the getLayersById function

* Req'd OL 4.3.1 + dropped mapbox-to-ol-style dep.

Dropped the custom getResolutionForZoom function in favor
   of the new ol.view.getResolutionForZoom function.

* Fixed for Legend defaultProps.

Fixed an issue where the legend was still rendering
internal text, causing the test to fail.

* More cleaning up of the ref/grouping.

Discovered a bug where the "diffs" for the update cycle
were not happening properly.  This was due to "type: ref" layers
being hydrated out of cycle.

While the "applyStyle" function can handle ref layers properly, the
layer that is referenced can be in a separate OL layer due to grouping.
When that happens applyStyle will not have the layers it needs to
resolve the "ref".

This refactors how layers were being hydrated from their refs.