Add ability to change the mode for drawing
it is possible with mapbox draw to have custom modes, this PR adds the
ability to assign them to the map and change them with the call to the
interactions.
there will be a mode for the interaction itself and one to switch to
after the interaction is done.
for example you want to draw new features, but after the feature is done
you don't want toi create another feature but rather change the newly
created one.